Async tests only run if Node >= v8
