+++ .travis.yml
@@ -2,5 +2,6 @@ language: node_js
 node_js:
 - '4.3.2'
 - '6.10.0'
+- '8.11.1'
 script:
 - npm run unit-tests
